django_app/requirements.txt
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
invoke==1.6.0
psycopg2==2.9.3
djangorestframework_simplejwt==5.2.0
djangorestframework-jwt==1.11.0
requests==2.28.2
django-rest-swagger==2.2.0
uWSGI==2.0.21